fix: reimport assets for build
