+++ public/assets/simulation/components/ac_source.png
