+++ public/assets/simulation/components/capacitor.png
