+++ public/assets/simulation/components/dc_source.png
