+++ public/assets/simulation/components/diode.png
