+++ public/assets/simulation/components/inductor.png
