+++ public/assets/simulation/components/pulse_source.png
