+++ public/assets/simulation/components/resistor.png
