+++ public/assets/simulation/tools/currentmeter.png
