+++ public/assets/simulation/tools/ohmmimeter.png
