+++ public/assets/simulation/tools/osciloscope.png
