+++ public/assets/simulation/tools/voltimeter.png
