+++ src/assets/component.ts
@@ -1,0 +1,17 @@
+import ac_source from "/assets/simulation/components/ac_source.png";
+import dc_source from "/assets/simulation/components/dc_source.png";
+import pulse_source from "/assets/simulation/components/pulse_source.png";
+import capacitor from "/assets/simulation/components/capacitor.png";
+import diode from "/assets/simulation/components/diode.png";
+import resistor from "/assets/simulation/components/resistor.png";
+import inductor from "/assets/simulation/components/inductor.png";
+
+export const components = {
+  ac_source,
+  dc_source,
+  pulse_source,
+  capacitor,
+  diode,
+  resistor,
+  inductor,
+};
