+++ src/assets/icons.ts
@@ -1,0 +1,7 @@
+import Alert from "/assets/icons/alert.png";
+import Eye from "/assets/icons/eye.png";
+import FacebookIcon from "/assets/icons/facebook-logo.png";
+import GoogleIcon from "/assets/icons/google-logo.png";
+import Hidden from "/assets/icons/hidden.png";
+
+export { Alert, Eye, FacebookIcon, GoogleIcon, Hidden };
