--- src/assets/icons/index.ts
@@ -1,7 +1,0 @@
-import Alert from "./alert.png";
-import Eye from "./eye.png";
-import FacebookIcon from "./facebook-logo.png";
-import GoogleIcon from "./google-logo.png";
-import Hidden from "./hidden.png";
-
-export { Alert, Eye, FacebookIcon, GoogleIcon, Hidden };
