--- src/assets/simulation/components/index.ts
@@ -1,17 +1,0 @@
-import resistor from "./resistor.png";
-import capacitor from "./capacitor.png";
-import dc_source from "./dc_source.png";
-import inductor from "./inductor.png";
-import diode from "./diode.png";
-import ac_source from "./ac_source.png";
-import pulse_source from "./pulse_source.png";
-
-export const components = {
-  resistor,
-  capacitor,
-  inductor,
-  diode,
-  dc_source,
-  ac_source,
-  pulse_source,
-};
