--- src/assets/simulation/tools/index.ts
@@ -1,11 +1,0 @@
-import voltimeter from "./voltimeter.png";
-import currentmeter from "./currentmeter.png";
-import ohmmimeter from "./ohmmimeter.png";
-import osciloscope from "./osciloscope.png";
-
-export const tools = {
-  voltimeter,
-  currentmeter,
-  ohmmimeter,
-  osciloscope,
-};
