+++ src/assets/tools.ts
@@ -1,0 +1,11 @@
+import currentmeter from "/assets/simulation/tools/currentmeter.png";
+import ohmmimeter from "/assets/simulation/tools/ohmmimeter.png";
+import voltimeter from "/assets/simulation/tools/voltimeter.png";
+import osciloscope from "/assets/simulation/tools/osciloscope.png";
+
+export const tools = {
+  currentmeter,
+  ohmmimeter,
+  voltimeter,
+  osciloscope,
+};
