+++ src/components/Icon/icons.ts
@@ -1,13 +1,13 @@
-import close from "../../assets/icons/close.svg";
-import edit from "../../assets/icons/edit.svg";
-import leftCurlyArrow from "../../assets/icons/left-curly-arrow.svg";
-import rightCurlyArrow from "../../assets/icons/right-curly-arrow.svg";
-import play from "../../assets/icons/play.svg";
-import print from "../../assets/icons/print.svg";
-import share from "../../assets/icons/share.svg";
-import trash from "../../assets/icons/trash.svg";
-import rotate from "../../assets/icons/rotate.svg";
-import pause from "../../assets/icons/pause.svg";
+import close from "/assets/icons/close.svg";
+import edit from "/assets/icons/edit.svg";
+import leftCurlyArrow from "/assets/icons/left-curly-arrow.svg";
+import rightCurlyArrow from "/assets/icons/right-curly-arrow.svg";
+import play from "/assets/icons/play.svg";
+import print from "/assets/icons/print.svg";
+import share from "/assets/icons/share.svg";
+import trash from "/assets/icons/trash.svg";
+import rotate from "/assets/icons/rotate.svg";
+import pause from "/assets/icons/pause.svg";
 
 export const icons = {
   close,
