+++ tsconfig.json
@@ -16,5 +16,5 @@
     "noEmit": true,
     "jsx": "react-jsx"
   },
-  "include": ["./src", "index.d.ts", "./public/wasm"]
+  "include": ["./src", "index.d.ts", "./public"]
 }
